Moving Event() to python - bugfix 4.
