+++ buildconfig/stubs/pygame/_common.pyi
@@ -1,5 +1,5 @@
 from os import PathLike
-from typing import IO, Callable, Tuple, Union, TypeVar
+from typing import IO, Callable, Tuple, Union, TypeVar, Dict, Any, Optional
 
 from typing_extensions import Literal as Literal, SupportsIndex as SupportsIndex
 from typing_extensions import Protocol
@@ -21,6 +21,25 @@ class Sequence(Protocol[_T]):
     """
     def __getitem__(self, __i: SupportsIndex) -> _T: ...
     def __len__(self) -> int: ...
+
+
+class EventLike(Protocol):
+    # __dict__: Dict[str, Any]
+    def __init__(
+        self, type: int, dict: Optional[Dict[str, Any]] = None, **kwargs: Any
+    ) -> None: ...
+    def __new__(cls, *args, **kwargs) -> "EventLike": ...
+    def __getattribute__(self, name: str) -> Any: ...
+    def __setattr__(self, name: str, value: Any) -> None: ...
+    def __delattr__(self, name: str) -> None: ...
+    def __int__(self) -> int: ...
+    def __bool__(self) -> bool: ...
+    def __eq__(self, other) -> bool: ...
+
+    @property
+    def type(self) -> int: ...
+    @property
+    def dict(self) -> Dict[str, Any]: ...
 
 # Right now, it isn't possible to annotate sizes (popular tools don't support it) but
 # when it is, the below types should be appropriately annotated
