+++ buildconfig/stubs/pygame/_event.pyi
@@ -3,9 +3,10 @@ from typing import (
     List,
     Optional,
     Union,
+    Type,
 )
 
-from ._common import Sequence
+from ._common import Sequence, EventLike
 
 _EventTypes = Union[int, Sequence[int]]
 
@@ -14,9 +15,9 @@ def get(
     eventtype: Optional[_EventTypes] = None,
     pump: Any = True,
     exclude: Optional[_EventTypes] = None,
-) -> List[Event]: ...
-def poll() -> Event: ...
-def wait(timeout: int = 0) -> Event: ...
+) -> List[EventLike]: ...
+def poll() -> EventLike: ...
+def wait(timeout: int = 0) -> EventLike: ...
 def peek(eventtype: Optional[_EventTypes] = None, pump: Any = True) -> bool: ...
 def clear(eventtype: Optional[_EventTypes] = None, pump: Any = True) -> None: ...
 def set_blocked(type: Optional[_EventTypes], /) -> None: ...
@@ -24,5 +25,5 @@ def set_allowed(type: Optional[_EventTypes], /) -> None: ...
 def get_blocked(type: _EventTypes, /) -> bool: ...
 def set_grab(grab: bool, /) -> None: ...
 def get_grab() -> bool: ...
-def post(event: Event, /) -> bool: ...
-def register_event_class(cls: type): ...
+def post(event: EventLike, /) -> bool: ...
+def register_event_class(cls: Type[EventLike]): ...
